# [C++] Optimize tagMap generation.