nb: I see no difference on my machine.

before:
```
C++:

Benchmark 1: ./main
Processing time (w/o IO): 17 ms
Processing time (w/o IO): 17 ms
Processing time (w/o IO): 16 ms
Processing time (w/o IO): 17 ms
Processing time (w/o IO): 16 ms
Processing time (w/o IO): 17 ms
Processing time (w/o IO): 16 ms
Processing time (w/o IO): 16 ms
Processing time (w/o IO): 16 ms
Processing time (w/o IO): 17 ms
Processing time (w/o IO): 16 ms
Processing time (w/o IO): 16 ms
Processing time (w/o IO): 17 ms
```

after:

```
C++:

Benchmark 1: ./main
Processing time (w/o IO): 17 ms
Processing time (w/o IO): 16 ms
Processing time (w/o IO): 16 ms
Processing time (w/o IO): 17 ms
Processing time (w/o IO): 17 ms
Processing time (w/o IO): 17 ms
Processing time (w/o IO): 16 ms
Processing time (w/o IO): 17 ms
Processing time (w/o IO): 17 ms
Processing time (w/o IO): 16 ms
Processing time (w/o IO): 17 ms
Processing time (w/o IO): 16 ms
Processing time (w/o IO): 17 ms
```

Results from the testing VM are below