@jinyus 2023-10-24
Main Branch:
```
C++ Concurrent | 18.08 ms | 220.00 ms | 1.88 s
````

This PR:
```
C++ Concurrent:

        Benchmark 1: ./main_con
        Processing time (w/o IO): 13 ms
        Processing time (w/o IO): 14 ms
        Processing time (w/o IO): 14 ms
        Processing time (w/o IO): 13 ms
        Processing time (w/o IO): 13 ms
        Processing time (w/o IO): 13 ms
        Processing time (w/o IO): 13 ms
        Processing time (w/o IO): 12 ms
        Processing time (w/o IO): 13 ms
        Processing time (w/o IO): 14 ms
        Processing time (w/o IO): 14 ms
        Processing time (w/o IO): 12 ms
        Processing time (w/o IO): 13 ms
          Time (mean ± σ):     227.4 ms ±   2.0 ms    [User: 231.3 ms, System: 29.7 ms]
          Range (min … max):   225.4 ms … 230.6 ms    10 runs
         
C++ Concurrent:

        Benchmark 1: ./main_con
        Processing time (w/o IO): 165 ms
        Processing time (w/o IO): 165 ms
        Processing time (w/o IO): 165 ms
          Time (mean ± σ):      1.035 s ±  0.001 s    [User: 1.374 s, System: 0.138 s]
          Range (min … max):    1.034 s …  1.036 s    2 runs
         
C++ Concurrent:

        Benchmark 1: ./main_con
        Processing time (w/o IO): 1403 ms
        Processing time (w/o IO): 1401 ms
        Processing time (w/o IO): 1401 ms
          Time (mean ± σ):      4.198 s ±  0.011 s    [User: 7.847 s, System: 0.330 s]
          Range (min … max):    4.190 s …  4.206 s    2 runs
```