@jinyus 2023-10-24
Single Threaded:

Main Branch:
`C++ | 24.85 ms | 339.00 ms | 2.92 s`

This PR:
```
C++:

        Benchmark 1: ./main
        Processing time (w/o IO): 24 ms
        Processing time (w/o IO): 24 ms
        Processing time (w/o IO): 24 ms
        Processing time (w/o IO): 24 ms
        Processing time (w/o IO): 24 ms
        Processing time (w/o IO): 24 ms
        Processing time (w/o IO): 24 ms
        Processing time (w/o IO): 24 ms
        Processing time (w/o IO): 24 ms
        Processing time (w/o IO): 24 ms
        Processing time (w/o IO): 24 ms
        Processing time (w/o IO): 24 ms
        Processing time (w/o IO): 24 ms
          Time (mean ± σ):     229.0 ms ±   2.0 ms    [User: 195.6 ms, System: 33.2 ms]
          Range (min … max):   226.6 ms … 232.3 ms    10 runs
         
C++:

        Benchmark 1: ./main
        Processing time (w/o IO): 343 ms
        Processing time (w/o IO): 343 ms
        Processing time (w/o IO): 343 ms
          Time (mean ± σ):      1.180 s ±  0.007 s    [User: 1.062 s, System: 0.112 s]
          Range (min … max):    1.175 s …  1.185 s    2 runs
         
C++:

        Benchmark 1: ./main
        Processing time (w/o IO): 3002 ms
        Processing time (w/o IO): 3006 ms
        Processing time (w/o IO): 3005 ms
          Time (mean ± σ):      5.588 s ±  0.013 s    [User: 5.163 s, System: 0.320 s]
          Range (min … max):    5.579 s …  5.597 s    2 runs
```