@jinyus 2023-10-24
Fresh run on main branch (to rule out noise):
```
C++:

        Benchmark 1: ./main
        Processing time (w/o IO): 25 ms
        Processing time (w/o IO): 25 ms
        Processing time (w/o IO): 25 ms
        Processing time (w/o IO): 25 ms
        Processing time (w/o IO): 25 ms
        Processing time (w/o IO): 25 ms
        Processing time (w/o IO): 25 ms
        Processing time (w/o IO): 25 ms
        Processing time (w/o IO): 25 ms
        Processing time (w/o IO): 25 ms
        Processing time (w/o IO): 25 ms
        Processing time (w/o IO): 25 ms
        Processing time (w/o IO): 25 ms
          Time (mean ± σ):     227.0 ms ±   1.7 ms    [User: 190.8 ms, System: 36.1 ms]
          Range (min … max):   224.1 ms … 229.8 ms    10 runs
         
C++:

        Benchmark 1: ./main
        Processing time (w/o IO): 338 ms
        Processing time (w/o IO): 339 ms
        Processing time (w/o IO): 339 ms
          Time (mean ± σ):      1.195 s ±  0.042 s    [User: 1.055 s, System: 0.112 s]
          Range (min … max):    1.166 s …  1.225 s    2 runs
         
C++:

        Benchmark 1: ./main
        Processing time (w/o IO): 2941 ms
        Processing time (w/o IO): 2929 ms
        Processing time (w/o IO): 2913 ms
          Time (mean ± σ):      5.496 s ±  0.004 s    [User: 5.089 s, System: 0.314 s]
          Range (min … max):    5.493 s …  5.499 s    2 runs
         
```